Create connect.php
In this example, the oci_connect function is used to connect to an Oracle database with the specified username, password, and database connection string (hostname and service name). The oci_close function is used to close the connection to the database. If the connection fails, an error message is displayed using the trigger_error function.

Note that you need to have the Oracle client libraries and the OCI8 extension installed on your server to use this code. You also need to have the correct connection information for your Oracle database, such as the hostname, service name, username, and password